requirements.txt
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
termcolor==2.4.0
typing-extensions==4.12.2
urllib3==2.1.0
